requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asyncstdlib-fw==3.13.2
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
botocore==1.38.41
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.4
decorator==4.4.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
eval_type_backport==0.2.2
exa-py==1.14.6
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-adk==1.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.96.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.19.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-core==0.3.65
langchain-google-genai==2.0.10
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
litellm==1.72.4
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.3
mdurl==0.1.2
mmh3==5.1.0
moviepy==1.0.3
msgpack==1.1.1
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.42.0
numpy==2.3.0
oauthlib==3.3.1
omegaconf==2.3.0
openai==1.86.0
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
posthog==5.3.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.9.10
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
tweepy==4.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
nest_asyncio
